+++ R/gMAP.R
@@ -347,27 +347,27 @@
 #'
 #' @template example-stop
 #' @export
-gMAP <- function (formula,
-                  family = gaussian,
-                  data,
-                  weights,
-                  offset,
-                  tau.strata,
-                  tau.dist=c("HalfNormal","TruncNormal","Uniform","Gamma","InvGamma","LogNormal","TruncCauchy","Exp", "Fixed"),
+gMAP <- function(formula,
+                 family = gaussian,
+                 data,
+                 weights,
+                 offset,
+                 tau.strata,
+                 tau.dist=c("HalfNormal","TruncNormal","Uniform","Gamma","InvGamma","LogNormal","TruncCauchy","Exp", "Fixed"),
                   tau.prior,
-                  tau.strata.pred=1,
-                  beta.prior,
-                  prior_PD=FALSE,
+                 tau.strata.pred=1,
+                 beta.prior,
+                 prior_PD=FALSE,
                   REdist=c("normal","t"),
-                  t.df=5,
-                  contrasts=NULL,
-                  iter=getOption("RBesT.MC.iter" , 6000),
-                  warmup=getOption("RBesT.MC.warmup", 2000),
-                  thin=getOption("RBesT.MC.thin", 4),
-                  init=getOption("RBesT.MC.init", 1),
-                  chains=getOption("RBesT.MC.chains", 4),
-                  cores=getOption("mc.cores", 1L)
-                  ) {
+                 t.df=5,
+                 contrasts=NULL,
+                 iter=getOption("RBesT.MC.iter" , 6000),
+                 warmup=getOption("RBesT.MC.warmup", 2000),
+                 thin=getOption("RBesT.MC.thin", 4),
+                 init=getOption("RBesT.MC.init", 1),
+                 chains=getOption("RBesT.MC.chains", 4),
+                 cores=getOption("mc.cores", 1L)
+                 ) {
     call <- match.call()
 
     if (is.character(family))